requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
arviz=0.15.1=pyhd8ed1ab_0
binutils=2.39=hdd6e379_1
binutils_impl_linux-64=2.39=he00db2b_1
binutils_linux-64=2.39=h5fc0e48_12
brotli=1.0.9=h166bdaf_8
brotli-bin=1.0.9=h166bdaf_8
brotlipy=0.7.0=py39hb9d737c_1005
bzip2=1.0.8=h7f98852_4
c-ares=1.18.1=h7f98852_0
c-compiler=1.3.0=h7f98852_0
ca-certificates=2023.08.22=h06a4308_0
cached-property=1.5.2=hd8ed1ab_1
cached_property=1.5.2=pyha770c72_1
cachetools=5.3.0=pyhd8ed1ab_0
certifi=2023.7.22=py39h06a4308_0
cffi=1.15.1=py39he91dace_3
charset-normalizer=3.1.0=pyhd8ed1ab_0
cloudpickle=2.2.1=py39h06a4308_0
contourpy=1.0.7=py39h4b4f3f3_0
cryptography=40.0.1=py39h079d5ae_0
cudatoolkit=11.8.0=h37601d7_11
cudnn=8.4.1.50=hed8a83a_0
cxx-compiler=1.3.0=h4bd325d_0
cycler=0.11.0=pyhd8ed1ab_0
decorator=5.1.1=pyhd8ed1ab_0
dill=0.3.6=pyhd8ed1ab_1
et_xmlfile=1.1.0=pyhd8ed1ab_0
fastprogress=1.0.3=pyhd8ed1ab_0
filelock=3.11.0=pyhd8ed1ab_0
fonttools=4.39.3=py39h72bdee0_0
freetype=2.12.1=hca18f0e_1
future=0.18.3=pyhd8ed1ab_0
gcc=9.5.0=h1fea6ba_12
gcc_impl_linux-64=9.5.0=h99780fb_19
gcc_linux-64=9.5.0=h4258300_12
gpy=1.10.0=py39hd257fcd_3
gxx=9.5.0=h1fea6ba_12
gxx_impl_linux-64=9.5.0=h99780fb_19
gxx_linux-64=9.5.0=h43f449f_12
h5netcdf=1.1.0=pyhd8ed1ab_1
h5py=3.8.0=nompi_py39h89bf01e_101
hdf5=1.14.0=nompi_hb72d44e_103
idna=3.4=pyhd8ed1ab_0
imbalanced-learn=0.10.1=pyhd8ed1ab_0
importlib-metadata=6.3.0=pyha770c72_0
importlib-resources=5.12.0=pyhd8ed1ab_0
importlib_resources=5.12.0=pyhd8ed1ab_0
jax=0.4.8=pyhd8ed1ab_0
jaxlib=0.4.7=cuda112py39h95e9e80_200
joblib=1.2.0=pyhd8ed1ab_0
kernel-headers_linux-64=2.6.32=he073ed8_15
keyutils=1.6.1=h166bdaf_0
kiwisolver=1.4.4=py39hf939315_1
krb5=1.20.1=h81ceb04_0
lcms2=2.15=haa2dc70_1
ld_impl_linux-64=2.39=hcc3a1bd_1
lerc=4.0.0=h27087fc_0
libabseil=20230125.0=cxx17_hcb278e6_1
libaec=1.0.6=hcb278e6_1
libblas=3.9.0=16_linux64_openblas
libbrotlicommon=1.0.9=h166bdaf_8
libbrotlidec=1.0.9=h166bdaf_8
libbrotlienc=1.0.9=h166bdaf_8
libcblas=3.9.0=16_linux64_openblas
libcurl=7.88.1=hdc1c0ab_1
libdeflate=1.18=h0b41bf4_0
libedit=3.1.20191231=he28a2e2_2
libev=4.33=h516909a_1
libffi=3.4.2=h7f98852_5
libgcc-devel_linux-64=9.5.0=h0a57e50_19
libgcc-ng=12.2.0=h65d4601_19
libgfortran-ng=12.2.0=h69a702a_19
libgfortran5=12.2.0=h337968e_19
libgomp=12.2.0=h65d4601_19
libgpuarray=0.7.6=h7f98852_1003
libgrpc=1.52.1=hcf146ea_1
libjpeg-turbo=2.1.5.1=h0b41bf4_0
liblapack=3.9.0=16_linux64_openblas
libllvm10=10.0.1=hbcb73fb_5
libnghttp2=1.52.0=h61bc06f_0
libnsl=2.0.0=h7f98852_0
libopenblas=0.3.21=pthreads_h78a6416_3
libpng=1.6.39=h753d276_0
libprotobuf=3.21.12=h3eb15da_0
libsanitizer=9.5.0=h2f262e1_19
libsqlite=3.40.0=h753d276_0
libssh2=1.10.0=hf14f497_3
libstdcxx-devel_linux-64=9.5.0=h0a57e50_19
libstdcxx-ng=12.2.0=h46fd767_19
libtiff=4.5.0=ha587672_6
libuuid=2.38.1=h0b41bf4_0
libwebp-base=1.3.0=h0b41bf4_0
libxcb=1.13=h7f98852_1004
libzlib=1.2.13=h166bdaf_4
llvmlite=0.36.0=py39h612dafd_4
mako=1.2.4=pyhd8ed1ab_0
markupsafe=2.1.2=py39h72bdee0_0
matplotlib-base=3.7.1=py39he190548_0
ml_dtypes=0.0.4=py39h7c7b5a8_0
munkres=1.1.4=pyh9f0ad1d_0
nccl=2.14.3.1=h0800d71_0
ncurses=6.3=h27087fc_1
nslr=0.1.3=pypi_0
numba=0.53.1=py39ha9443f7_0
numpy=1.24.2=py39h7360e5f_0
numpy-indexed=0.3.5=py_1
openjpeg=2.5.0=hfec8fc6_2
openpyxl=3.1.1=py39h72bdee0_0
openssl=3.1.0=h0b41bf4_0
opt_einsum=3.3.0=pyhd8ed1ab_1
packaging=23.0=pyhd8ed1ab_0
pandas=1.5.3=py39h2ad29b5_1
paramz=0.9.5=py_0
patsy=0.5.3=pyhd8ed1ab_0
pillow=9.5.0=py39h7207d5c_0
pip=23.0.1=pyhd8ed1ab_0
platformdirs=3.2.0=pyhd8ed1ab_0
pooch=1.7.0=pyha770c72_3
pthread-stubs=0.4=h36c2ea0_1001
pycparser=2.21=pyhd8ed1ab_0
pygpu=0.7.6=py39hce5d2b2_1003
pymc3=3.11.2=py39ha563b64_4
pyopenssl=23.1.1=pyhd8ed1ab_0
pyparsing=3.0.9=pyhd8ed1ab_0
pysocks=1.7.1=pyha2e5f31_6
pystable=0.2.3=pypi_0
python=3.9.16=h2782a2a_0_cpython
python-dateutil=2.8.2=pyhd8ed1ab_0
python_abi=3.9=3_cp39
pytz=2023.3=pyhd8ed1ab_0
re2=2023.02.02=hcb278e6_0
readline=8.2=h8228510_1
requests=2.28.2=pyhd8ed1ab_1
scikit-learn=1.2.2=py39hd189fd4_1
scipy=1.10.1=py39h7360e5f_0
seaborn=0.12.2=hd8ed1ab_0
seaborn-base=0.12.2=pyhd8ed1ab_0
semver=3.0.0=pyhd8ed1ab_0
setuptools=57.5.0=pypi_0
shap=0.42.1=py39h1128e8f_0
six=1.16.0=pyh6c4a22f_0
slicer=0.0.7=pyhd3eb1b0_0
statsmodels=0.13.5=py39h2ae25f5_2
sysroot_linux-64=2.12=he073ed8_15
tbb=2020.3=hfd86e86_0
theano-pymc=1.1.2=py39he80948d_0
threadpoolctl=3.1.0=pyh8a188c0_0
tk=8.6.12=h27826a3_0
tqdm=4.65.0=py39hb070fc8_0
typing-extensions=4.5.0=hd8ed1ab_0
typing_extensions=4.5.0=pyha770c72_0
tzdata=2023c=h71feb2d_0
unicodedata2=15.0.0=py39hb9d737c_0
urllib3=1.26.15=pyhd8ed1ab_0
vonmisesmixtures=1.0.0=pypi_0
wheel=0.40.0=pyhd8ed1ab_0
xarray=2023.3.0=pyhd8ed1ab_0
xarray-einstats=0.5.1=pyhd8ed1ab_0
xorg-libxau=1.0.9=h7f98852_0
xorg-libxdmcp=1.1.3=h7f98852_0
xz=5.2.6=h166bdaf_0
zipp=3.15.0=pyhd8ed1ab_0
zlib=1.2.13=h166bdaf_4
zstd=1.5.2=h3eb15da_6
